requirements.txt
tensorflow-cpu==1.15.0
nameko==3.0.0rc6
